fix(planning_test_utils): fix build error `-werror=sign-compare` (#7086)
fix(planning_test_utils): fix build error

+++ planning/planning_test_utils/test/test_mock_data_parser.cpp
@@ -80,11 +80,12 @@ TEST(ParseFunctions, CompleteYAMLTest)
 
   // Test parsing of segments
   std::vector<LaneletSegment> segments = parse_segments(config["segments"]);
-  ASSERT_EQ(segments.size(), 1);  // Assuming only one segment in the provided YAML for this test
+  ASSERT_EQ(
+    segments.size(), uint64_t(1));  // Assuming only one segment in the provided YAML for this test
 
   const auto & segment0 = segments[0];
   EXPECT_EQ(segment0.preferred_primitive.id, 11);
-  EXPECT_EQ(segment0.primitives.size(), 2);
+  EXPECT_EQ(segment0.primitives.size(), uint64_t(2));
   EXPECT_EQ(segment0.primitives[0].id, 22);
   EXPECT_EQ(segment0.primitives[0].primitive_type, "lane");
   EXPECT_EQ(segment0.primitives[1].id, 33);
@@ -118,10 +119,10 @@ TEST(ParseFunction, CompleteFromFilename)
 
   ASSERT_EQ(
     lanelet_route.segments.size(),
-    2);  // Assuming only one segment in the provided YAML for this test
+    uint64_t(2));  // Assuming only one segment in the provided YAML for this test
   const auto & segment1 = lanelet_route.segments[1];
   EXPECT_EQ(segment1.preferred_primitive.id, 44);
-  EXPECT_EQ(segment1.primitives.size(), 4);
+  EXPECT_EQ(segment1.primitives.size(), uint64_t(4));
   EXPECT_EQ(segment1.primitives[0].id, 55);
   EXPECT_EQ(segment1.primitives[0].primitive_type, "lane");
   EXPECT_EQ(segment1.primitives[1].id, 66);
